ConfigInstanceBuilder use Converters to load
