+++ implementation/src/main/java/io/smallrye/config/ConfigInstanceBuilder.java
@@ -211,12 +211,24 @@ public interface ConfigInstanceBuilder<I> {
      * Globally registers a {@link org.eclipse.microprofile.config.spi.Converter} to be used by the
      * {@link io.smallrye.config.ConfigInstanceBuilder} to convert configuration interface member types.
      *
-     * @param type the class of the type to convert
-     * @param converter the converter instance that can convert to the type
+     * @param type the class of the type to convert (must not be {@code null})
+     * @param converter the converter instance that can convert to the type (must not be {@code null})
      * @param <T> the type to convert
      */
     static <T> void registerConverter(Class<T> type, Converter<T> converter) {
         ConfigInstanceBuilderImpl.CONVERTERS.put(type, converter);
+    }
+
+    /**
+     * Resolve a {@link org.eclipse.microprofile.config.spi.Converter} for {@code type}, throwing if none can be found.
+     *
+     * @param type the type to be produced by the converter (must not be {@code null})
+     * @return an instance of the {@link Converter} for the specified type (not {@code null})
+     * @param <T> the type to convert
+     * @throws IllegalArgumentException if no converter is registered for the given type
+     */
+    static <T> Converter<T> requireConverter(final Class<T> type) {
+        return Converters.requireConverter(ConfigInstanceBuilderImpl.CONVERTERS, type);
     }
 
     /**
