+++ implementation/src/test/java/io/smallrye/config/ObjectCreatorTest.java
@@ -16,8 +16,8 @@ import java.util.function.Supplier;
 
 import org.junit.jupiter.api.Test;
 
-import io.smallrye.config.ConfigInstanceBuilderImpl.MapWithDefault;
 import io.smallrye.config.ConfigMapping.NamingStrategy;
+import io.smallrye.config.ConfigMappingContext.ObjectCreator.MapWithDefault;
 
 public class ObjectCreatorTest {
     @Test
